// apps/app/src/components/types.ts
import type { LinkProps } from "next/link";

interface User {
    name: string
    email: string
    avatar: string
}

interface Team {
    name: string
    logo: React.ElementType
    plan: string
}

interface BaseNavItem {
    title: string
    badge?: string
    icon?: React.ElementType
}

type NavLink = BaseNavItem & {
    url: LinkProps["href"]
    items?: never
}

type NavCollapsible = BaseNavItem & {
    items: (BaseNavItem & { url: LinkProps["href"] })[]
    url?: never
}

type NavItem = NavCollapsible | NavLink

interface NavGroup {
    title: string
    items: NavItem[]
}

interface SidebarData {
    user: User
    teams: Team[]
    navGroups: NavGroup[]
}

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink }